local env set up
