+++ package.json
@@ -18,11 +18,11 @@
     "@expo/vector-icons": "14.0.4",
     "@react-navigation/native": "6.0.2",
     "@supabase/supabase-js": "2.46.1",
-    "expo": "51.0.39",
+    "expo": "^52.0.11",
     "expo-constants": "~16.0.2",
     "expo-font": "12.0.10",
     "expo-linking": "~6.3.1",
-    "expo-router": "3.5.24",
+    "expo-router": "^3.1.0",
     "expo-splash-screen": "0.27.7",
     "expo-status-bar": "~1.12.1",
     "expo-system-ui": "~3.0.7",
